@@ -6,7 +6,7 @@ repository: https://github.com/imujtaba8488/im_animations.git
 issue_tracker: https://github.com/imujtaba8488/im_animations/issues
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
